renaming UCF directory to DFB, since it now contains TALL
