+++ base_miner/DFB/config/tall.yaml
