+++ base_miner/DFB/detectors/base_detector.py
