+++ base_miner/DFB/detectors/tall_detector.py
